fixed: slow draw caused by wrong Camera::getViewableArea() being called.  Removed deprecated slow function.
todo: should change Display::draw() to pull data down rather than have Game push.

+++ trunk/zomgame/ZGame/renderable.h
@@ -12,17 +12,17 @@ class Renderable {
 	protected:
 		short color;
 		string desc;
-		int displayChar;
+		long displayChar;
 		string name;
 		
 	public:
 		virtual short getColor() {return color;}
 		virtual string getDescription() {return desc;}
-		virtual int getDisplayChar() {return displayChar;}
+		virtual long getDisplayChar() {return displayChar;}
 		virtual string getName() {return name;}
 		virtual void setName(string newName) {name = newName;}
 		virtual void setDescription(string newDesc) {desc = newDesc;}
-		virtual void setDisplayChar(int newDispChar) {displayChar = newDispChar;}
+		virtual void setDisplayChar(long newDispChar) {displayChar = newDispChar;}
 		virtual void setColor(short newColor) {color = newColor;}
 
 };
